+++ MiniTycoonPiekarnia/Models/Ingredient.cs
@@ -3,7 +3,7 @@
 public class Ingredient
 {
     public string Name { get; set; } = string.Empty;
-    public int Quantity { get; set; } = 0;
+    public decimal Quantity { get; set; } = 0;
     public decimal PurchasePrice { get; set; } = 0m;
 
     public string IconFile => $"{Normalize(Name)}.png";
